Refactor NewsletterForm to use centralized schema and types for form validation
